+++ src/serial/plugin-SiPixelClusterizer/gpu_calib_pixel.jl
@@ -25,7 +25,7 @@ const v_calto_electron_gain_L1::Float32 = 50       # L1:   49.6 +- 2.6
 const v_calto_electron_offset::Float32 = -60       # L2-4: -60 +- 130
 const v_calto_electron_offset_L1::Float32 = -670   # L1: -670 +- 200
 
-function calib_digis(is_run_2::Bool, id::Vector{Int16}, x::Vector{Int16}, y::Vector{Int16}, adc::Vector{Int32}, ped::SiPixelGainForHLTonGPU, num_elements::Integer, module_start::Vector{UInt32}, n_clusters_in_module::Vector{UInt32}, clus_module_start::Vector{UInt32})
+function calib_digis(is_run_2::Bool, id::Vector{UInt16}, x::Vector{UInt16}, y::Vector{UInt16}, adc::Vector{UInt16}, ped::SiPixelGainForHLTonGPU, num_elements::Integer, module_start::Vector{UInt32}, n_clusters_in_module::Vector{UInt32}, clus_module_start::Vector{UInt32})
     first = 0
     
     # zero for next kernels
